Made Navbar js file
