+++ src/pages/Home.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import NavBar from "./NavBar";
 
 function checkUserLogin(currentUser, navigate) {
   if (currentUser === null || currentUser.trim() === "") {
@@ -17,41 +18,20 @@ function Home() {
     checkUserLogin(currentUser, navigate)
   })
 
-  function logOut(){
+  function logOut() {
     localStorage.setItem("CUR_USER", "")
     navigate("/login")
   }
 
-  function storage(){
+  function storage() {
     navigate("/storage")
   }
 
   return (
-
-    <nav className="navbar navbar-expand-lg bg-body-tertiary">
-      <div className="container-fluid">
-        <a className="navbar-brand" href="/#">Pantry Pal</a>
-        <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-          <span className="navbar-toggler-icon"></span>
-        </button>
-        <div className="collapse navbar-collapse" id="navbarNav">
-          <ul className="navbar-nav">
-            <li className="nav-item">
-              <a className="nav-link active" aria-current="page" href="/#" >Home</a>
-            </li>
-            <li className="nav-item">
-              <a className="nav-link"  onClick={storage}>Add Inventory</a>
-            </li>
-            <li className="nav-item">
-              <a className="nav-link"  onClick={logOut}>Setting</a>
-            </li>
-            <li className="nav-item">
-              <a className="nav-link"   onClick={logOut}>Logout</a>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </nav>
+    <div>
+      <NavBar />
+      
+    </div>
 
 
   )
